moved action client handling to a generic class
